Separate LocalRemotePIP and Modal from MobilePane
